+++ src/config.js
@@ -115,7 +115,7 @@ var options = mergeConfig({
     // jscs:enable maximumLineLength
     report: {
       disable: envs('AZK_REPORT_DISABLE', false),
-      url    : envs('AZK_REPORT_URL', "http://api.azk.io/report/A2ObXDzlG6fOJC803qBzDjde2YxygZY4"),
+      url    : envs('AZK_REPORT_URL', "http://api.azk.io/report/N2ekMXt7OlwscGV8V40o4WMI6gzT31WM"),
       // TODO: Adding support to filters
       // filters: envs('AZK_REPORT_FITLERS', "*"),
     },
@@ -194,9 +194,6 @@ var options = mergeConfig({
       projectId: envs('AZK_KEEN_PROJECT_ID', '5526968d672e6c5a0d0ebec6'),
       writeKey : envs('AZK_KEEN_WRITE_KEY', '5dbce13e376070e36eec0c7dd1e7f42e49f39b4db041f208054617863832309c14a797409e12d976630c3a4b479004f26b362506e82a46dd54df0c977a7378da280c05ae733c97abb445f58abb56ae15f561ac9ad774cea12c3ad8628d896c39f6e702f6b035541fc1a562997cb05768'),
       // jscs:enable maximumLineLength
-    },
-    report: {
-      url: envs('AZK_REPORT_URL', "http://force-stage.azk.io/report/A2ObXDzlG6fOJC803qBzDjde2YxygZY4"),
     },
     urls: {
       github: {