+++ src/libexec/package-tools/pack.sh
@@ -44,6 +44,11 @@ abs_dir() {
   if [ -z "$link" ]; then pwd; else abs_dir $link; fi
 }
 
+#Clean environment
+unset BUILD_DEB BUILD_RPM BUILD_MAC SECRET_KEY RELEASE_CHANNEL \
+      NO_VERSION NO_MAKE NO_CLEAN_LINUX CLEAN_REPO CLEAN_REPO_ARGS \
+      NO_AGENT NO_TEST PUBLISH NO_TAG VERBOSE
+
 export AZK_ROOT_PATH=`cd \`abs_dir ${BASH_SOURCE:-$0}\`/../../..; pwd`
 export AZK_BUILD_TOOLS_PATH=${AZK_ROOT_PATH}/src/libexec/package-tools
 
@@ -76,7 +81,8 @@ while [[ $# -gt 0 ]]; do
     --no-linux-clean )
       NO_CLEAN_LINUX=true;;
     --clean-repo )
-      CLEAN_REPO='--clean-repo';;
+      CLEAN_REPO=true
+      CLEAN_REPO_ARGS='--clean-repo';;
     --no-agent )
       NO_AGENT=true;;
     --no-test )
@@ -93,17 +99,17 @@ while [[ $# -gt 0 ]]; do
    esac
 done
 
-[[ -z "${RELEASE_CHANNEL}" ]] && RELEASE_CHANNEL='stable'
-VERSION_SUFFIX="-${RELEASE_CHANNEL}"
+PACKAGE_SUFFIX="-${RELEASE_CHANNEL}"
 
 case $RELEASE_CHANNEL in
-  nightly )
-    AZK_BALANCER_IP=192.168.60.4;;
+  "" | stable )
+    AZK_BALANCER_IP=192.168.62.4
+    RELEASE_CHANNEL=stable
+    PACKAGE_SUFFIX=;;
   rc )
     AZK_BALANCER_IP=192.168.61.4;;
-  stable )
-    AZK_BALANCER_IP=192.168.62.4
-    VERSION_SUFFIX=;;
+  nightly )
+    AZK_BALANCER_IP=192.168.60.4;;
   * ) echo >&2 "Invalid release channel: ${RELEASE_CHANNEL}." && exit 2;;
 esac
 
@@ -113,37 +119,37 @@ if [[ -z "${BUILD_DEB}" ]] && [[ -z "${BUILD_RPM}" ]] && [[ -z "${BUILD_MAC}" ]]
   BUILD_MAC=true
 fi
 
-if [[ ! -z "${BUILD_DEB}" ]] || [[ ! -z "${BUILD_RPM}" ]]; then
-  [[ ! -e $SECRET_KEY ]] && echo >&2 "Please inform an valid GPG key." && exit 3
+if [[ ${BUILD_DEB} == true ]] || [[ ${BUILD_RPM} == true ]]; then
+  [[ ! -e "${SECRET_KEY}" ]] && echo >&2 "Please inform a valid GPG key." && exit 3
 fi
 
-bump_version() {
+calculate_azk_version() {
   VERSION_NUMBER=$( cat package.json | grep -e "version" | cut -d' ' -f4 | sed -n 's/\"//p' | sed -n 's/\"//p' | sed -n 's/,//p' | sed s/-.*// )
 
   if [[ $RELEASE_CHANNEL != 'stable' ]]; then
     RELEASE_COUNTER=$(curl -s https://api.github.com/repos/azukiapp/azk/tags | grep 'name' | \
       grep -c "${VERSION_NUMBER}-${RELEASE_CHANNEL}" | awk '{print $1 + 1}' )
     RELEASE_DATE=$( date +%Y%m%d )
-    VERSION_SUFFIX_NO_META="${VERSION_SUFFIX}.${RELEASE_COUNTER}"
+    VERSION_SUFFIX_NO_META="${PACKAGE_SUFFIX}.${RELEASE_COUNTER}"
     VERSION_SUFFIX="${VERSION_SUFFIX_NO_META}+${RELEASE_DATE}"
   fi
 
   VERSION="${VERSION_NUMBER}${VERSION_SUFFIX}"
   VERSION_NO_META="${VERSION_NUMBER}${VERSION_SUFFIX_NO_META}"
+}
 
-  if [[ $NO_VERSION != true ]]; then
-    files=( package.json npm-shrinkwrap.json )
-    for f in "${files[@]}"; do
-      VERSION_LINE_NUMBER=`cat ${f} | grep -n "version" | head -1 | cut -d ":" -f1`
-      rm -Rf ${f}r # Avoiding conflicts
-      sed -ir "${VERSION_LINE_NUMBER}s/\([[:digit:]]*\.[[:digit:]]*\.[[:digit:]]*\)[^\"]*/\1${VERSION_SUFFIX}/" ${f}
-      rm -Rf ${f}r
-      git add ${f}
-    done
-    git commit -m "Bumping version to azk v${VERSION_NO_META}"
+bump_version() {
+  files=( package.json npm-shrinkwrap.json )
+  for f in "${files[@]}"; do
+    VERSION_LINE_NUMBER=`cat ${f} | grep -n "version" | head -1 | cut -d ":" -f1`
+    rm -Rf ${f}r # Avoiding conflicts
+    sed -ir "${VERSION_LINE_NUMBER}s/\([[:digit:]]*\.[[:digit:]]*\.[[:digit:]]*\)[^\"]*/\1${VERSION_SUFFIX}/" ${f}
+    rm -Rf ${f}r
+    git add ${f}
+  done
+  git commit -m "Bumping version to azk v${VERSION_NO_META}"
 
-    echo "Version bumped to v${VERSION}."
-  fi
+  echo "Version bumped to v${VERSION}."
 }
 
 make_tag() {
@@ -202,11 +208,23 @@ start_agent() {
   export AZK_NAMESPACE="${RELEASE_CHANNEL}.release.azk.io"
   export AZK_BALANCER_HOST=$AZK_NAMESPACE
   export AZK_VM_MEMORY=768
+  export AZK_DNS_PORT=53
+
+  RESOLVER_FILE_NAME="/etc/resolver/${AZK_NAMESPACE}"
+  RESOLVER_FILE_CONTENT="nameserver ${AZK_BALANCER_IP}.${AZK_DNS_PORT}"
+
+  if [ "$(cat ${RESOLVER_FILE_NAME})" != "${RESOLVER_FILE_CONTENT}" ]; then
+    echo "${RESOLVER_FILE_CONTENT}" \
+      | sudo tee ${RESOLVER_FILE_NAME} \
+      > /dev/null 2>&1
+  fi
+
+  azk config set terms_of_use.accepted 1 > /dev/null 2>&1
 
   azk agent stop
   sleep 3
 
-  ./bin/azk agent start --no-daemon > $AZK_AGENT_LOG_FILE 2>&1 &
+  azk agent start --no-daemon > $AZK_AGENT_LOG_FILE 2>&1 &
   AGENT_PID="$!"
   tail -F $AZK_AGENT_LOG_FILE &
   TAIL_PID="$!"
@@ -235,8 +253,9 @@ source .dependencies
 
 LINUX_BUILD_WAS_EXECUTED=false
 
-step_run "Bumping version" --exit bump_version
+calculate_azk_version
 
+[[ $NO_VERSION != true ]] && step_run "Bumping version" --exit bump_version
 [[ $NO_MAKE != true ]]    && step_run "Running make" --exit run_make
 [[ $NO_AGENT != true ]]   && step_run "Starting agent" --exit start_agent
 [[ $NO_TEST != true ]]    && step_run "Preparing testing env" setup_test && TEST_ARGS=$TEST_DIR
@@ -254,8 +273,10 @@ if [[ $BUILD_DEB == true ]]; then
   (
     set -e
 
-    [[ ! -z "${CLEAN_REPO}" ]] && step_run "Cleaning current aptly repo" azk shell --shell=/bin/sh package -c "rm -Rf /azk/aptly/*"
-    [[ ! -z "${CLEAN_REPO}" ]] && step_run "Cleaning environment" rm -Rf package/deb package/public
+    if [[ ${CLEAN_REPO} == true ]]; then
+      step_run "Cleaning current aptly repo" azk shell --shell=/bin/sh package -c "rm -Rf /azk/aptly/*"
+      step_run "Cleaning environment" rm -Rf package/deb package/public
+    fi
 
     step_run "Downloading libnss-resolver" \
     mkdir -p package/deb \
@@ -268,7 +289,9 @@ if [[ $BUILD_DEB == true ]]; then
       EXTRA_FLAGS="LINUX_CLEAN="
     fi
 
-    step_run "Creating deb packages" make package_deb ${EXTRA_FLAGS}
+    step_run "Creating deb packages" \
+    rm -rf /azk/aptly/public/pool/main/a/azk${PACKAGE_SUFFIX}/azk${PACKAGE_SUFFIX}_${VERSION}_amd64.deb \
+    && make package_deb ${EXTRA_FLAGS}
 
     UBUNTU_VERSIONS=( "ubuntu12:precise" "ubuntu14:trusty" "ubuntu15:wily" )
     for UBUNTU_VERSION in "${UBUNTU_VERSIONS[@]}"; do
@@ -277,7 +300,7 @@ if [[ $BUILD_DEB == true ]]; then
       UBUNTU_VERSION_CODENAME="${UBUNTU_VERSION##*:}"
 
       step_run "Generating ${UBUNTU_VERSION_NUMBER} repository" \
-        linux_generate ubuntu ${LIBNSS_RESOLVER_VERSION} ${UBUNTU_VERSION_CODENAME} ${CLEAN_REPO}
+        linux_generate ubuntu ${LIBNSS_RESOLVER_VERSION} ${UBUNTU_VERSION_CODENAME} ${CLEAN_REPO_ARGS}
 
       if [[ $NO_TEST != true ]]; then
         step_run "Testing ${UBUNTU_VERSION_NUMBER} repository" \
@@ -302,7 +325,7 @@ if [[ $BUILD_RPM == true ]]; then
   (
     set -e
 
-    [[ ! -z "${CLEAN_REPO}" ]] && step_run "Cleaning environment" rm -Rf package/rpm package/fedora20 package/fedora23
+    [[ ${CLEAN_REPO} == true ]] && step_run "Cleaning environment" rm -Rf package/rpm package/fedora20 package/fedora23
 
     step_run "Downloading libnss-resolver" \
     mkdir -p package/rpm \
@@ -319,9 +342,10 @@ if [[ $BUILD_RPM == true ]]; then
     FEDORA_VERSIONS=( "fedora20" "fedora23" )
     for FEDORA_VERSION in "${FEDORA_VERSIONS[@]}"; do
       step_run "Generating ${FEDORA_VERSION} repository" \
-        linux_generate fedora ${FEDORA_VERSION} ${CLEAN_REPO}
+        linux_generate fedora ${FEDORA_VERSION} ${CLEAN_REPO_ARGS}
       if [[ $NO_TEST != true ]]; then
-        step "Testing ${FEDORA_VERSION} repository" ${AZK_BUILD_TOOLS_PATH}/test.sh ${FEDORA_VERSION} ${TEST_ARGS}
+        step_run "Testing ${FEDORA_VERSION} repository" \
+          ${AZK_BUILD_TOOLS_PATH}/test.sh ${FEDORA_VERSION} ${TEST_ARGS}
       fi
     done