+++ src/system/run.js
@@ -449,20 +449,26 @@ var Run = {
 
   _clean_sync_folder(system, host_folder) {
     return async(this, function* () {
-      var uid_gid;
+      var local_user = config('agent:vm:user');
+      var uid, gid;
       if (config('agent:requires_vm')) {
-        uid_gid = `\$(id -u ${config('agent:vm:user')}):\$(id -g ${config('agent:vm:user')})`;
+        uid = `\$(id -u ${local_user})`;
+        gid = `\$(id -g ${local_user})`;
       } else {
-        uid_gid = `${process.getuid()}:${process.getuid()}`;
+        uid = `${process.getuid()}`;
+        gid = uid;
       }
 
       var mounted_sync_folders = '/sync_folders';
       var current_sync_folder = path.join(mounted_sync_folders, system.manifest.namespace, system.name, host_folder);
 
+      var find_exec = `-exec chown ${uid}:${gid} '{}' \\;`;
+      var find_args = `${current_sync_folder} \\( -not -user ${uid} -or -not -group ${gid} \\) ${find_exec}`;
+
       // Script to fix sync folder
       var script = [
         `mkdir -p ${current_sync_folder}`,
-        `chown -R ${uid_gid} ${mounted_sync_folders}`
+        `find ${find_args}`,
       ].join(" && ");
 
       // Docker params